将列表抽到新 ViewModel 中
