+++ src/WindowDebugger/Views/MainViewModel.cs
@@ -1,142 +1,24 @@
-using System.Collections.Immutable;
-using System.Diagnostics;
 using Avalonia.Collections;
 using ReactiveUI;
 using WindowDebugger.Services.NativeWindows;
-using WindowDebugger.Services.NativeWindows.Linux;
-using WindowDebugger.Services.NativeWindows.Windows;
-using WindowDebugger.Utils;
 
 namespace WindowDebugger.Views;
 
 public record MainViewModel : ReactiveRecord
 {
-    private readonly NativeWindowCollectionManager _nativeWindowCollectionManager;
-    private string? _searchText;
-    private bool _includingInvisibleWindow;
-    private bool _includingEmptyTitleWindow;
-    private bool _includingChildWindow;
-    private bool _includingMessageOnlyWindow;
-
-    public MainViewModel()
-    {
-        _nativeWindowCollectionManager = new();
-    }
+    public WindowListViewModel WindowList { get; }
 
     public AvaloniaList<NativeTreeNode> NativeTree { get; } = [];
 
-    public string? SearchText
+    public MainViewModel()
     {
-        get => _searchText;
-        set => this.RaiseAndSetWithAction(ref _searchText, value, ReloadWindows);
-    }
-
-    public bool IncludingInvisibleWindow
-    {
-        get => _includingInvisibleWindow;
-        set => this.RaiseAndSetWithAction(ref _includingInvisibleWindow, value, ReloadWindows);
-    }
-
-    public bool IncludingEmptyTitleWindow
-    {
-        get => _includingEmptyTitleWindow;
-        set => this.RaiseAndSetWithAction(ref _includingEmptyTitleWindow, value, ReloadWindows);
-    }
-
-    public bool IncludingChildWindow
-    {
-        get => _includingChildWindow;
-        set => this.RaiseAndSetWithAction(ref _includingChildWindow, value, ReloadWindows);
-    }
-
-    public bool IncludingMessageOnlyWindow
-    {
-        get => _includingMessageOnlyWindow;
-        set => this.RaiseAndSetWithAction(ref _includingMessageOnlyWindow, value, ReloadWindows);
+        WindowList = new WindowListViewModel(this);
     }
 
     public void ReloadWindows()
     {
-        var nativeWindows = _nativeWindowCollectionManager.FindWindows(new WindowSearchingFilter(SearchText)
-        {
-            IncludingInvisibleWindow = IncludingInvisibleWindow,
-            IncludingEmptyTitleWindow = IncludingEmptyTitleWindow,
-            IncludingChildWindow = IncludingChildWindow,
-            IncludingMessageOnlyWindow = IncludingMessageOnlyWindow,
-        });
-
-        var tree = BuildTree(nativeWindows);
-
+        var tree = WindowList.ReloadWindows();
         NativeTree.Clear();
         NativeTree.AddRange(tree);
-    }
-
-    private IEnumerable<NativeTreeNode> BuildTree(ImmutableArray<NativeWindowModel> nativeWindows)
-    {
-        if (OperatingSystem.IsLinux())
-        {
-            var processWindowDictionary = GroupByProcess<LinuxNativeWindowModel>(nativeWindows);
-            return processWindowDictionary
-                .Select(x => new NativeProcessNode(x.Key)
-                {
-                    ProcessName = TryGetProcessName(x.Key),
-                    Windows = [..x.Value.Select(ConvertModelToNode)],
-                });
-        }
-
-        if (OperatingSystem.IsWindows())
-        {
-            var processWindowDictionary = GroupByProcess<WindowsNativeWindowModel>(nativeWindows);
-            return processWindowDictionary
-                .Select(x => new NativeProcessNode(x.Key)
-                {
-                    ProcessName = TryGetProcessName(x.Key),
-                    Windows =
-                    [
-                        ..x.Value.Select<WindowsNativeWindowModel, NativeWindowNode>(w => new WindowsNativeWindowNode(w)
-                        {
-                            ChildWindows = [],
-                        }),
-                    ],
-                });
-        }
-
-        throw new PlatformNotSupportedException();
-    }
-
-    private static NativeWindowNode ConvertModelToNode(LinuxNativeWindowModel model)
-    {
-        return new LinuxNativeWindowNode(model)
-        {
-            ChildWindows = [..model.GetChildren().Select(ConvertModelToNode)],
-        };
-    }
-
-    private static Dictionary<int, List<T>> GroupByProcess<T>(ImmutableArray<NativeWindowModel> nativeWindows)
-        where T : NativeWindowModel
-    {
-        var processWindowDictionary = new Dictionary<int, List<T>>();
-        foreach (var window in nativeWindows.OfType<T>())
-        {
-            if (!processWindowDictionary.TryGetValue(window.ProcessId, out var processWindows))
-            {
-                processWindows = [];
-                processWindowDictionary[window.ProcessId] = processWindows;
-            }
-            processWindows.Add(window);
-        }
-        return processWindowDictionary;
-    }
-
-    private static string? TryGetProcessName(int pid)
-    {
-        try
-        {
-            return Process.GetProcessById(pid).ProcessName;
-        }
-        catch
-        {
-            return null;
-        }
     }
 }
